Added new Exception TooManyScrollContextsException and registered it with ElasticsearchException
